requirements.txt
streamlit>=1.28.0
requests>=2.31.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymongo==4.6.1
google-genai==0.2.2
pydantic==2.9.2
pydantic-settings==2.6.1
PyYAML==6.0.1
langgraph==0.2.0
python-dotenv==1.0.0
httpx==0.26.0

